+++ resources/release-notes.md
@@ -5,3 +5,5 @@
 ## Software
 
 - (All) Add support for plotting multiple datasets with the same variables in `mag.graphics.chart.Stackedplot`
+- (IMAP) Add health checks for PROCSTAT HK, including single bit error, missed and rejected ITF frames
+- (IMAP) Do not run secondary voltages and currents health checks on EM